pyproject.toml
[project]
name = "aiops-tools"
version = "0.1.0"
description = "Agent Tools Management Platform - 管理 AI Agent 可调用的工具"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "your@email.com" }]
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    # Database
    "sqlalchemy>=2.0.0",
    "sqlmodel>=0.0.14",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Redis & Celery
    "redis>=5.0.0",
    "celery[redis]>=5.3.0",
    # Validation & Serialization
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "jsonschema>=4.20.0",
    # Utils
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "tenacity>=8.2.0",
    # Operations Tools Dependencies
    "kubernetes>=28.1.0",
    "psycopg2-binary>=2.9.9",
    "mysql-connector-python>=8.3.0",
    "boto3>=1.34.0",
    "greenlet>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aiops_tools"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
